feat: configure eslint, prettier and husky
