+++ .prettierrc.js
@@ -1,6 +1,7 @@
 module.exports = {
-  bracketSpacing: false,
-  jsxBracketSameLine: true,
+  printWidth: 100,
   singleQuote: true,
+  bracketSpacing: true,
+  jsxBracketSameLine: true,
   trailingComma: 'all',
 };
