+++ src/components/Fact/Fact.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import {object} from 'prop-types';
-import {Text, View} from 'react-native';
+import { object } from 'prop-types';
+import { Text, View } from 'react-native';
 
 import styles from './Fact.styles';
 
-const Fact = ({fact: {_id, text, upvotes}}) => (
+const Fact = ({ fact: { _id, text, upvotes } }) => (
   <View key={_id} style={styles.container}>
     <Text style={styles.text}>{text}</Text>
     <View>
